backend/requirements.txt
Flask
Pillow
tensorflow
scikit-learn
opencv-python
scikit-image